+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 ## Change log
 ----------------------
 
+Version 1.2
+-------------
+
+CHANGED:
+
+- update of parent version to 5
+- added new prefixes for unique constraints and foreign keys in enum DatabasePrefix
+- update of ent-api dependency to version 1.1
+- replaced hibernate with javax.persistence-api dependency 
+
 Version 1.1
 -------------